+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This version REQUIRES the MSCLoader!
 ### Changes
 
 - MOP has been ported to MSCLoader by Piotrulos
+- 'mop help' now is shown alphabetically
 
 ### Bug Fixes